requirements.txt
pytz==2023.3
astronomer-cosmos==1.8.2
apache-airflow-providers-postgres==5.6.0
dbt-core==1.6.1
dbt-postgres==1.6.1
